Added: 1. Form validation for add test component added including insurance that start date cannot be after the end date. 2. New unit tests added for entity and repository layers.

+++ backendAA/src/test/java/com/pgault04/repositories/TestTestQuestionRepo.java
@@ -1,6 +1,3 @@
-/**
- * 
- */
 package com.pgault04.repositories;
 
 import static org.junit.Assert.*;
@@ -22,8 +19,8 @@ import com.pgault04.repositories.TestQuestionRepo;
 
 
 /**
- * @author paulgault
- *
+ * @author Paul Gault 40126005
+ * @since November 2018
  */
 @Sql("/tests.sql")
 @Transactional
@@ -42,117 +39,67 @@ public class TestTestQuestionRepo {
 
 	private TestQuestion testQuestion;
 
-	/**
-	 * @throws java.lang.Exception
-	 */
 	@Before
 	public void setUp() throws Exception {
-
 		testQuestion = new TestQuestion(TEST_ID_IN_DB, QUESTION_ID_IN_DB);
-
 	}
 
-	/**
-	 * Test method for Row Count
-	 */
 	@Test
 	public void testRowCount() {
-
-		int rowCountBefore = testQuestionRepo.rowCount().intValue();
-		
+		int rowCountBefore = testQuestionRepo.rowCount();
 		// Inserts one test to table
 		testQuestionRepo.insert(testQuestion);
-
 		// Checks one value is registered as in the table
-		assertTrue(testQuestionRepo.rowCount().intValue() > rowCountBefore);
-
+		assertTrue(testQuestionRepo.rowCount() > rowCountBefore);
 	}
 
-	/**
-	 * TestQuestion method for
-	 * {@link com.pgault04.assessmentassistant.repositories.TestQuestionsRepo#insert(com.pgault04.assessmentassistant.entities.TestQuestions)}.
-	 */
 	@Test
 	public void testTestQuestionInsert() {
-
 		// Inserts one testQuestion to table
 		TestQuestion returnedTestQuestion = testQuestionRepo.insert(testQuestion);
-
 		List<TestQuestion> testQuestions = testQuestionRepo
 				.selectByTestQuestionID(returnedTestQuestion.getTestQuestionID());
-
 		assertEquals(1, testQuestions.size());
-
 		// Updates the testQuestion in the table
 		returnedTestQuestion.setQuestionID(SECOND_QUESTION_ID_IN_DB);
-
 		// Inserts one testQuestion to table
 		testQuestionRepo.insert(returnedTestQuestion);
-
 		testQuestions = testQuestionRepo.selectByTestQuestionID(returnedTestQuestion.getTestQuestionID());
-
 		assertEquals(SECOND_QUESTION_ID_IN_DB, testQuestions.get(0).getQuestionID().intValue());
-
 	}
 
-	/**
-	 * TestQuestion method for
-	 * {@link com.pgault04.assessmentassistant.repositories.TestQuestionsRepo#selectByTestQuestionsID(java.lang.Long)}.
-	 */
 	@Test
 	public void testTestQuestionSelectByTestQuestionID() {
 		// Inserts one testQuestion to table
 		TestQuestion returnedTestQuestion = testQuestionRepo.insert(testQuestion);
-
 		List<TestQuestion> testQuestions = testQuestionRepo
 				.selectByTestQuestionID(returnedTestQuestion.getTestQuestionID());
-
 		assertEquals(1, testQuestions.size());
 	}
 
-	/**
-	 * TestQuestion method for
-	 * {@link com.pgault04.assessmentassistant.repositories.TestQuestionsRepo#selectByUsername(java.lang.String)}.
-	 */
 	@Test
 	public void testTestQuestionSelectByTestID() {
 		// Inserts one testQuestion to table
 		testQuestionRepo.insert(testQuestion);
-
 		List<TestQuestion> testQuestions = testQuestionRepo.selectByTestID(testQuestion.getTestID());
-
 		assertTrue(testQuestions.size() > 0);
 	}
 
-	/**
-	 * TestQuestion method for
-	 * {@link com.pgault04.assessmentassistant.repositories.TestQuestionsRepo#selectByUsername(java.lang.String)}.
-	 */
 	@Test
 	public void testTestQuestionSelectByQuestionID() {
 		// Inserts one testQuestion to table
 		testQuestionRepo.insert(testQuestion);
-
 		List<TestQuestion> testQuestions = testQuestionRepo.selectByQuestionID(testQuestion.getQuestionID());
-
 		assertTrue(testQuestions.size() > 0);
 	}
 
-	/**
-	 * TestQuestion method for
-	 * {@link com.pgault04.assessmentassistant.repositories.TestQuestionsRepo#delete(java.lang.Long)}.
-	 */
 	@Test
 	public void testTestQuestionDelete() {
 		// Inserts one testQuestion to table
 		TestQuestion returnedTestQuestions = testQuestionRepo.insert(testQuestion);
-
 		testQuestionRepo.delete(returnedTestQuestions.getTestQuestionID());
-
 		List<TestQuestion> testQuestions = testQuestionRepo
 				.selectByTestQuestionID(returnedTestQuestions.getTestQuestionID());
-
 		assertEquals(0, testQuestions.size());
 	}
-
 }
